+++ core/CMakeLists.txt
@@ -12,7 +12,7 @@ find_package(argparse REQUIRED)
 
 add_library(core INTERFACE)
 target_include_directories(core INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
-target_link_libraries(core INTERFACE cppzmq capnp capnp_schemas kj)
+target_link_libraries(core INTERFACE cppzmq capnpc capnp capnp_schemas kj)
 target_compile_features(core INTERFACE)
 
 # only for intellisense purposes
@@ -20,7 +20,7 @@ file(GLOB_RECURSE CORE_HEADERS "*.h" "*.hpp" "*.hxx")
 if(CORE_HEADERS)
     add_library(core_intellisense EXCLUDE_FROM_ALL ${CORE_HEADERS})
     target_include_directories(core_intellisense PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
-    target_link_libraries(core_intellisense PRIVATE cppzmq capnp capnp_schemas kj)
+    target_link_libraries(core_intellisense PRIVATE cppzmq capnpc capnp capnp_schemas kj)
     target_compile_features(core_intellisense PRIVATE cxx_std_17)
     set_target_properties(core_intellisense PROPERTIES
         LINKER_LANGUAGE CXX
