+++ rerun/CMakeLists.txt
@@ -15,6 +15,6 @@ add_executable(rerun viewer.cpp)
 add_executable(rerun_comprobation service/comprobation.cpp)
 add_executable(rerun_render_registry_example render_registry_example.cpp)
 
-target_link_libraries(rerun rerun_sdk core cppzmq capnp capnp_schemas kj ${PCL_LIBRARIES})
-target_link_libraries(rerun_comprobation core cppzmq capnp capnp_schemas kj)
-target_link_libraries(rerun_render_registry_example rerun_sdk core cppzmq capnp capnp_schemas kj)
+target_link_libraries(rerun rerun_sdk core cppzmq capnpc capnp capnp_schemas kj ${PCL_LIBRARIES})
+target_link_libraries(rerun_comprobation core cppzmq capnpc capnp capnp_schemas kj)
+target_link_libraries(rerun_render_registry_example rerun_sdk core cppzmq capnpc capnp capnp_schemas kj)
